backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pillow>=11.0.0
pytesseract>=0.3.13
nltk>=3.9.0
python-multipart>=0.0.18
